+++ packages/logic-core/src/data/rules/musicGridRule.ts
@@ -104,8 +104,12 @@ export default class MusicGridRule
     return MusicGridRule.SEARCH_VARIANTS;
   }
 
-  public validateGrid(_grid: GridData): RuleState {
-    return { state: State.Incomplete };
+  public validateGrid(grid: GridData): RuleState {
+    if (grid.getTileCount(true, false, Color.Gray) > 0) {
+      return { state: State.Incomplete };
+    } else {
+      return { state: State.Satisfied };
+    }
   }
 
   public onSetGrid(