+++ packages/logic-core/src/data/symbols/directionLinkerSymbol.ts
@@ -105,11 +105,15 @@ export default class DirectionLinkerSymbol extends Symbol {
     return DirectionLinkerSymbol.EXAMPLE_GRID;
   }
 
-  private deltaCoordinate(c: Position, direction: Direction): Position {
-    return {
-      x: c.x + DirectionLinkerSymbol.directionDeltas[direction].dx,
-      y: c.y + DirectionLinkerSymbol.directionDeltas[direction].dy,
-    };
+  private deltaCoordinate(
+    c: Position,
+    direction: Direction,
+    grid: GridData
+  ): Position {
+    return grid.toArrayCoordinates(
+      c.x + DirectionLinkerSymbol.directionDeltas[direction].dx,
+      c.y + DirectionLinkerSymbol.directionDeltas[direction].dy
+    );
   }
 
   public validateSymbol(grid: GridData): State {
@@ -175,8 +179,12 @@ export default class DirectionLinkerSymbol extends Symbol {
         const directions = Object.keys(this.linkedDirections) as Direction[];
         for (const direction of directions) {
           const newTurtle: Turtle = {
-            pos1: this.deltaCoordinate(pos1, direction),
-            pos2: this.deltaCoordinate(pos2, this.linkedDirections[direction]),
+            pos1: this.deltaCoordinate(pos1, direction, grid),
+            pos2: this.deltaCoordinate(
+              pos2,
+              this.linkedDirections[direction],
+              grid
+            ),
             color1: baseColor1,
             color2: baseColor2,
           };