+++ packages/logic-core/src/index.ts
@@ -9,6 +9,7 @@ import Configurable from './data/configurable.js';
 import { CachedAccess, allEqual, array, directionToRotation, escape, isSameEdge, maxBy, minBy, move, orientationToRotation, resize, unescape } from './data/dataHelper.js';
 import { isEventHandler } from './data/events/eventHelper.js';
 import { handlesFinalValidation } from './data/events/onFinalValidation.js';
+import { handlesGetTile } from './data/events/onGetTile.js';
 import { handlesGridChange } from './data/events/onGridChange.js';
 import { handlesGridResize } from './data/events/onGridResize.js';
 import { handlesSetGrid, invokeSetGrid } from './data/events/onSetGrid.js';
@@ -18,7 +19,7 @@ import GridData, { NEIGHBOR_OFFSETS } from './data/grid.js';
 import GridConnections from './data/gridConnections.js';
 import GridZones from './data/gridZones.js';
 import Instruction from './data/instruction.js';
-import { COMPARISONS, Color, Comparison, DIRECTIONS, Direction, MajorRule, Mode, ORIENTATIONS, Orientation, State, directionToggle, orientationToggle } from './data/primitives.js';
+import { COMPARISONS, Color, Comparison, DIRECTIONS, Direction, MajorRule, Mode, ORIENTATIONS, Orientation, State, WRAPPINGS, Wrapping, directionToggle, orientationToggle } from './data/primitives.js';
 import { MetadataSchema, PuzzleSchema } from './data/puzzle.js';
 import BanPatternRule from './data/rules/banPatternRule.js';
 import CellCountPerZoneRule from './data/rules/cellCountPerZoneRule.js';
@@ -41,6 +42,7 @@ import SameShapeRule from './data/rules/sameShapeRule.js';
 import SymbolsPerRegionRule from './data/rules/symbolsPerRegionRule.js';
 import UndercluedRule from './data/rules/undercluedRule.js';
 import UniqueShapeRule from './data/rules/uniqueShapeRule.js';
+import WrapAroundRule from './data/rules/wrapAroundRule.js';
 import { Serializer } from './data/serializer/allSerializers.js';
 import { Compressor } from './data/serializer/compressor/allCompressors.js';
 import CompressorBase from './data/serializer/compressor/compressorBase.js';
@@ -125,6 +127,7 @@ export {
   unescape,
   isEventHandler,
   handlesFinalValidation,
+  handlesGetTile,
   handlesGridChange,
   handlesGridResize,
   handlesSetGrid,
@@ -146,6 +149,8 @@ export {
   ORIENTATIONS,
   Orientation,
   State,
+  WRAPPINGS,
+  Wrapping,
   directionToggle,
   orientationToggle,
   MetadataSchema,
@@ -172,6 +177,7 @@ export {
   SymbolsPerRegionRule,
   UndercluedRule,
   UniqueShapeRule,
+  WrapAroundRule,
   Serializer,
   Compressor,
   CompressorBase,